Add dedicated project for ontology export
