+++ platforms/snips-nlu-ontology-kotlin/settings.gradle
@@ -16,3 +16,4 @@ include 'services:webservice'
 */
 
 rootProject.name = 'snips-nlu-ontology'
+include ':snips-nlu-ontology-export'
