+++ platforms/snips-nlu-ontology-kotlin/snips-nlu-ontology-export/src/main/kotlin/com.sun.jna/JnaUtils.kt
@@ -1,0 +1,4 @@
+package com.sun.jna
+
+// NativeString is package private...
+fun String.toJnaPointer(encoding: String) = NativeString(this, encoding).pointer
